Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Pradeep,
> 
> Before upgrade why are we disabling NSR ?. As per issu procedure NSR should be enabled before ISSU (https://www.juniper.net/documentation/en_US/junos/topics/task/installation/issu-upgrading.html) . Will you please clarify this ?

We are not using the ISSU. We need to disable the NSR because we upgrade RE by RE. If both RE's are not having the same version then it will not work as expected. 

This is the field recomendation to follow this procedure.

Patch-set: 1
Reviewer: Gerrit User 843 <843@c72cd27d-1743-4a55-84a4-cd13b8400c52>